Separated power function out of utils
