+++ role-admin/src/main/java/org/rfcx/guardian/admin/comms/swm/SwmDispatchService.java
@@ -115,7 +115,7 @@ public class SwmDispatchService extends Service {
 									Log.e(logTag, "SWM Send Failure (Consecutive Failures: " + app.swmUtils.consecutiveDeliveryFailureCount + ")...");
 									if (app.swmUtils.consecutiveDeliveryFailureCount >= SwmUtils.powerCycleAfterThisManyConsecutiveDeliveryFailures) {
 										//app.swmUtils.setPower(false);
-										app.swmUtils.setPower(true);
+										app.swmUtils.getPower().setPower(true);
 										app.swmUtils.consecutiveDeliveryFailureCount = 0;
 										break;
 									}
